% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeViews.R
\name{closeViews}
\alias{closeViews}
\title{this function uses the dataframe from getRStudioOpenDocIDs to identify tabs opened by
View() as well as untitled tabs, and faciliates closing them}
\usage{
closeViews(keepUntitledFiles = T)
}
\description{
this function uses the dataframe from getRStudioOpenDocIDs to identify tabs opened by
View() as well as untitled tabs, and faciliates closing them
}
