% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mysti.R
\name{combine_mysti_obs}
\alias{combine_mysti_obs}
\title{Combine Mysticetus observations
Combine Mysticetus observations for multiple days over a survey}
\usage{
combine_mysti_obs(data.source = "cemore", year, month, day)
}
\arguments{
\item{day}{}
}
\description{
Combine Mysticetus observations
Combine Mysticetus observations for multiple days over a survey
}
\examples{
##Not run: combine_mysti_obs(year = 2021, month = "09",day = 20)
##End(Not run)
}
