% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mysti.R
\name{combine_mysti_tracks}
\alias{combine_mysti_tracks}
\title{Combine mysticetus tracks
Combine mysticetus tracklines for multiple days over a survey}
\usage{
combine_mysti_tracks(year, month, day)
}
\arguments{
\item{day}{}
}
\description{
Combine mysticetus tracks
Combine mysticetus tracklines for multiple days over a survey
}
\examples{
## Not run:
combine_mysti_tracks(2021, "09", 20)
## End(Not run)
}
