% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_bb}
\alias{create_bb}
\title{create_bb
Make a bounding box to clip plots by or create magnified inset}
\usage{
create_bb(xmin, xmax, ymin, ymax)
}
\arguments{
\item{ymax}{}
}
\value{

}
\description{
create_bb
Make a bounding box to clip plots by or create magnified inset
}
\examples{
bb <- create_bb(-123.4, -123.1, 48.32, 48.5)
cropped_sf <- st_intersection(sf, bb)
}
