% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lines}
\alias{get_lines}
\title{get_lines
Bind together list of survey transect lines brought in as gpx (readGPX) into one
linestring object. Used with tidy_transects() to import multiple lines in
.gpx file.}
\usage{
get_lines(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
get_lines
Bind together list of survey transect lines brought in as gpx (readGPX) into one
linestring object. Used with tidy_transects() to import multiple lines in
.gpx file.
}
