% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_obs_data}
\alias{get_obs_data}
\title{get_obs_data}
\usage{
get_obs_data(year, month, data.source = "cemore", vessel = "MB")
}
\arguments{
\item{data_folder}{}
}
\description{
get_obs_data
}
\examples{
data <- get_obs_data(data_folder = "sept2020/"); saveRDS(data, "survey_data/cemore_survey_data_oct2020.rds")
}
