% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeViews.R
\name{getRStudioOpenDocIDs}
\alias{getRStudioOpenDocIDs}
\title{code largely taken from https://github.com/stevenjwest/projectmanagr/blob/master/R/RStudio-functions.R,
but now works on windows (within Rstudio projects as well as with standalone instances of RStudio);
and it returns a dataframe with some information about each of the tabs}
\usage{
getRStudioOpenDocIDs(showDets = F)
}
\description{
code largely taken from https://github.com/stevenjwest/projectmanagr/blob/master/R/RStudio-functions.R,
but now works on windows (within Rstudio projects as well as with standalone instances of RStudio);
and it returns a dataframe with some information about each of the tabs
}
