% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_transects}
\alias{tidy_transects}
\title{Tidy transects
Read in gpx file of planned survey transects (as either route or track) and convert to sf LINESTRING object}
\usage{
tidy_transects(
  file_name,
  dir = NULL,
  type = "route",
  depth_cutoff_object = NULL
)
}
\arguments{
\item{route}{true/false whether the object is a route (= linestring object)}
}
\description{
Tidy transects
Read in gpx file of planned survey transects (as either route or track) and convert to sf LINESTRING object
}
\examples{
## Not run:: test <- tidy_transects(file_name = "18km_full_2.gpx", iteration = 2, route = FALSE)
}
